Fix command deploy commands
